move to single worker node-pool
this commit changes our distribution of concourse workers in a way to
have more web nodes that are half the size (vertically) than before, as
we've never been really fully utilizing their resources anyway.

it moves us from having 2 worker node pools of vertically large
machines, to a single pool with more nodes that are half the size that
they previously were (16vcpu 32GBram 1TBdisk --> 8vcpu 16GBram
750GBdisk).

the reasoning for not performing a "divide by 2" on `disk` is that
that's a resource that's still very highly utilized mostly in terms of
IOPS, but eventually, in terms of space too.

to achieve that, the `cluster` terraform module was updated to use a map
of node pools (rather than a list) so that we can update the set of
pools without needing to manually perform terraform surgery.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>