view kibana at tracing.concourse-ci.org
oauth2-proxy secrets live in lastpass, so I moved the chart to `with-creds`.

It was nice to discover that so many helm charts have a top-level `ingress:`
field that plays nicely with nginx.

The only problem was making sure that the load balancer would be deployed at the
IP that terraform provisioned for the `tracing.concourse-ci.org` domain - this
is why there is a separate nginx ingress controller in the jaeger namespace.

Signed-off-by: Jamie Klassen <cklassen@pivotal.io>