initial scaffolding
doesn't do much yet, just has a pipeline which actually just consumes a
'greenhouse' branch of the 'hush-house' repo - there may be some joining
of houses in the future in the interest of preserving git history. (i.e.
we might be better off just renaming the hush-house repo to greenpeace.)

the bootstrap/ folder sets up a buckets and a service account with
permissions only for the bucket. the service account credentials must be
fetched manually; they cannot really be terraformed, as you are only
able to acquire the key once after creation.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>