add google storage bucket for vault storage backend
Replace the cloudsql database with a GCS bucket. There are a few reasons
for this change:

1. Using the Postgres backend forces us to use a fork of the vault-helm
chart
2. The Postgres backend requires us to manually create DB tables, which
isn't super easy with terraform

The GCS bucket that's created has versioning enabled, but only keeps the
latest 3 versions - this is to avoid the problem described in
sethvargo/vault-on-gke#13, and solved in https://github.com/sethvargo/vault-on-gke/commit/f40dc7598ea5412485e67350a0054381c7de5e5b.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>