wip: add dispatcher cluster
In order to continuously deploy the production cluster, we can't be
deploying from the production cluster (even after bootstrapping, where
the production cluster won't exist). That's because terraform needs to
run somewhere - if we mess with concourse/the cluster while running
terraform, we may end up in a bad state where the .tfstate file isn't
accurate and the deployment fails.

The dispatcher cluster is separate from production - it's only purpose
is to deploy other clusters.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Bishoy Youssef <byoussef@pivotal.io>