split terraform modules into environments/dependencies
this way, we can more intelligently trigger on changes to the
environments' terraform modules. just triggering on changes to the
environment module itself isn't sufficient since changes may take place
in the modules that the environments depend on.

note that we haven't yet updated the triggers - this commit just
restructures the modules to make it easier to do so

Signed-off-by: Esteban Foronda <eforonda@vmware.com>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>