add service account for k8s_topgun cluster
the eventual goal is to get rid of basic auth for our k8s clusters (see
concourse/infrastructure#12). this commit adds a service account + key
that only has access to the cluster that we can use in CI to generate
the kube_config dynamically (e.g. using the gcloud CLI)

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Esteban Foronda <eforonda@vmware.com>