start stress worker nodes at 17
I suspect the reason why the terraform job keeps failing with connection
refused is because it is timing out before the cluster can fully come
up. I tried setting a higher timeout but there might be a bug in the
helm_release of terraform that makes it not respect the timeout. So
instead I'm going to try to initially provision a high number of nodes
for the workers so that it does not need to autoscale from 1 -> 17 nodes
for the workers. I think this is the reason why it takes so long from
looking at the stress cluster logs and seeing that the autoscaler was
running during the time between the job failing and then once the
autoscaler finished the cluster came up.

Signed-off-by: Clara Fu <fclara@vmware.com>