remove hashicorp/template provider
it's been deprecated for a long time and we can't even manage the
modules because the provider isn't available for darwin/arm64.

I had to manually remove all instances of hashicorp/template from the
state files as well.

Also just realized that we use workspaces to manage which state file
you're using. This is a wild footgun that we setup for ourselves and is
not the correct way to use workspaces at all. I might try migrating
these at a future date.

Signed-off-by: Taylor Silva <dev@taydev.net>